message

Module message

Re-exports

pub use crate::queue_name::QueueNameComponents;

Structs

Message

MessageList

A list of messages with an O(1) list overhead; no additional memory per-message is required to track this list. However, a given Message can only belong to one instance of such a list at a time.

MessageListIter

WeakMessage