Struct message::message::Message

pub struct Message { /* private fields */ }

Implementations

impl Message

pub fn new_dirty( id: SpoolId, sender: EnvelopeAddress, recipient: EnvelopeAddress, meta: Value, data: Arc<Box<[u8]>>, ) -> Result<Self>

Create a new message with the supplied data. The message meta and data are marked as dirty

pub fn weak(&self) -> WeakMessage

pub fn new_from_spool(id: SpoolId, metadata: Vec<u8>) -> Result<Self>

Helper for creating a message based on spool enumeration. Given a spool id and the serialized metadata blob, returns a message holding the deserialized version of that metadata.

pub async fn new_with_id(id: SpoolId) -> Result<Self>

pub fn get_num_attempts(&self) -> u16

pub fn set_num_attempts(&self, num_attempts: u16)

pub fn increment_num_attempts(&self)

pub fn set_scheduling(&self, scheduling: Option<Scheduling>) -> Result<()>

pub fn get_due(&self) -> Option<DateTime<Utc>>

pub async fn delay_with_jitter(&self, limit: i64) -> Result<()>

pub async fn delay_by(&self, duration: Duration) -> Result<()>

pub async fn delay_by_and_jitter(&self, duration: Duration) -> Result<()>

Delay by requested duration, and add up to 1 minute of jitter

pub async fn set_due(&self, due: Option<DateTime<Utc>>) -> Result<()>

pub fn set_force_sync(&self, force: bool)

pub fn needs_save(&self) -> bool

pub async fn save(&self) -> Result<()>

pub async fn save_to( &self, meta_spool: &(dyn Spool + Send + Sync), data_spool: &(dyn Spool + Send + Sync), ) -> Result<()>

pub fn id(&self) -> &SpoolId

pub fn shrink(&self) -> Result<bool>

pub fn sender(&self) -> Result<EnvelopeAddress>

pub fn set_sender(&self, sender: EnvelopeAddress) -> Result<()>

pub fn recipient(&self) -> Result<EnvelopeAddress>

pub fn set_recipient(&self, recipient: EnvelopeAddress) -> Result<()>

pub fn is_meta_loaded(&self) -> bool

pub fn is_data_loaded(&self) -> bool

pub async fn load_meta_if_needed(&self) -> Result<()>

pub async fn load_data_if_needed(&self) -> Result<()>

pub async fn load_meta(&self) -> Result<()>

pub async fn load_meta_from( &self, meta_spool: &(dyn Spool + Send + Sync), ) -> Result<()>

pub async fn load_data(&self) -> Result<()>

pub async fn load_data_from( &self, data_spool: &(dyn Spool + Send + Sync), ) -> Result<()>

pub fn assign_data(&self, data: Vec<u8>)

pub fn get_data(&self) -> Arc<Box<[u8]>>

pub fn set_meta<S: AsRef<str>, V: Into<Value>>( &self, key: S, value: V, ) -> Result<()>

pub fn get_meta_string<S: Index + Display + Copy>( &self, key: S, ) -> Result<Option<String>>

Retrieve key as a String.

pub fn get_meta_obj(&self) -> Result<Value>

pub fn get_meta<S: Index>(&self, key: S) -> Result<Value>

pub fn age(&self, now: DateTime<Utc>) -> Duration

pub fn get_queue_name(&self) -> Result<String>

pub async fn dkim_verify(&self) -> Result<Vec<AuthenticationResult>>

pub fn parse_rfc3464(&self) -> Result<Option<Report>>

pub fn parse_rfc5965(&self) -> Result<Option<ARFReport>>

pub fn prepend_header(&self, name: Option<&str>, value: &str)

pub fn append_header(&self, name: Option<&str>, value: &str)

pub fn get_address_header( &self, header_name: &str, ) -> Result<Option<HeaderAddressList>>

pub fn get_first_named_header_value(&self, name: &str) -> Result<Option<String>>

pub fn get_all_named_header_values(&self, name: &str) -> Result<Vec<String>>

pub fn get_all_headers(&self) -> Result<Vec<(String, String)>>

pub fn retain_headers<F: FnMut(&Header<'_>) -> bool>( &self, func: F, ) -> Result<()>

pub fn remove_first_named_header(&self, name: &str) -> Result<()>

pub fn import_x_headers(&self, names: Vec<String>) -> Result<()>

pub fn remove_x_headers(&self, names: Vec<String>) -> Result<()>

pub fn remove_all_named_headers(&self, name: &str) -> Result<()>

pub fn dkim_sign(&self, signer: &Signer) -> Result<()>

pub fn import_scheduling_header( &self, header_name: &str, remove: bool, ) -> Result<()>

pub fn append_text_plain(&self, content: &str) -> Result<bool>

pub fn append_text_html(&self, content: &str) -> Result<bool>

pub fn check_fix_conformance( &self, check: MessageConformance, fix: MessageConformance, ) -> Result<()>