message::message

Struct Message

pub struct Message { /* private fields */ }

Implementations

impl Message

pub fn new_dirty( id: SpoolId, sender: EnvelopeAddress, recipient: Vec<EnvelopeAddress>, meta: Value, data: Arc<Box<[u8]>>, ) -> Result<Self>

Create a new message with the supplied data. The message meta and data are marked as dirty

pub fn weak(&self) -> WeakMessage

pub fn new_from_spool(id: SpoolId, metadata: Vec<u8>) -> Result<Self>

Helper for creating a message based on spool enumeration. Given a spool id and the serialized metadata blob, returns a message holding the deserialized version of that metadata.

pub async fn new_with_id(id: SpoolId) -> Result<Self>

pub fn get_num_attempts(&self) -> u16

pub fn set_num_attempts(&self, num_attempts: u16)

pub fn increment_num_attempts(&self)

pub async fn set_scheduling( &self, scheduling: Option<Scheduling>, ) -> Result<Option<Scheduling>>

pub async fn get_scheduling(&self) -> Result<Option<Scheduling>>

pub fn get_due(&self) -> Option<DateTime<Utc>>

pub async fn delay_with_jitter( &self, limit: i64, ) -> Result<Option<DateTime<Utc>>>

pub async fn delay_by( &self, duration: Duration, ) -> Result<Option<DateTime<Utc>>>

pub async fn delay_by_and_jitter( &self, duration: Duration, ) -> Result<Option<DateTime<Utc>>>

Delay by requested duration, and add up to 1 minute of jitter

pub async fn set_due( &self, due: Option<DateTime<Utc>>, ) -> Result<Option<DateTime<Utc>>>

pub fn set_force_sync(&self, force: bool)

pub fn needs_save(&self) -> bool

pub async fn save(&self, deadline: Option<Instant>) -> Result<()>

pub async fn save_to( &self, meta_spool: &(dyn Spool + Send + Sync), data_spool: &(dyn Spool + Send + Sync), deadline: Option<Instant>, ) -> Result<()>

pub fn id(&self) -> &SpoolId

pub async fn save_and_shrink(&self) -> Result<bool>

Save the data+meta if needed, then release both

pub async fn save_and_shrink_data(&self) -> Result<bool>

Save the data+meta if needed, then release just the data

pub fn shrink_data(&self) -> Result<bool>

pub fn shrink(&self) -> Result<bool>

pub async fn sender(&self) -> Result<EnvelopeAddress>

pub async fn set_sender(&self, sender: EnvelopeAddress) -> Result<()>

pub async fn recipient(&self) -> Result<EnvelopeAddress>

👎Deprecated: use recipient_list or first_recipient instead

pub async fn first_recipient(&self) -> Result<EnvelopeAddress>

pub async fn recipient_list(&self) -> Result<Vec<EnvelopeAddress>>

pub async fn recipient_list_string(&self) -> Result<Vec<String>>

pub async fn set_recipient(&self, recipient: EnvelopeAddress) -> Result<()>

👎Deprecated: use set_recipient_list instead

pub async fn set_recipient_list( &self, recipient: Vec<EnvelopeAddress>, ) -> Result<()>

pub fn is_meta_loaded(&self) -> bool

pub fn is_data_loaded(&self) -> bool

pub async fn load_meta_if_needed(&self) -> Result<()>

pub async fn data(&self) -> Result<Arc<Box<[u8]>>>

pub async fn load_meta(&self) -> Result<()>

pub async fn load_data(&self) -> Result<Arc<Box<[u8]>>>

pub fn assign_data(&self, data: Vec<u8>)

pub fn get_data_maybe_not_loaded(&self) -> Arc<Box<[u8]>>

pub async fn set_meta<S: AsRef<str>, V: Into<Value>>( &self, key: S, value: V, ) -> Result<()>

pub async fn unset_meta<S: AsRef<str>>(&self, key: S) -> Result<()>

pub async fn get_meta_string<S: Index + Display + Copy>( &self, key: S, ) -> Result<Option<String>>

Retrieve key as a String.

pub async fn get_meta_obj(&self) -> Result<Value>

pub async fn get_meta<S: Index>(&self, key: S) -> Result<Value>

pub fn age(&self, now: DateTime<Utc>) -> Duration

pub async fn get_queue_name(&self) -> Result<String>

pub async fn arc_verify( &self, opt_resolver_name: Option<String>, ) -> Result<AuthenticationResult>

pub async fn arc_seal( &self, signer: Signer, auth_results: AuthenticationResults, opt_resolver_name: Option<String>, ) -> Result<()>

pub async fn dkim_verify( &self, opt_resolver_name: Option<String>, ) -> Result<Vec<AuthenticationResult>>

pub async fn parse(&self) -> Result<MimePart<'static>>

Parses the content into an owned MimePart. Changes to that MimePart are NOT reflected in the underlying message; you must re-assign the message data if you wish to modify the message content.

pub async fn parse_rfc3464(&self) -> Result<Option<Report>>

pub async fn parse_rfc5965(&self) -> Result<Option<ARFReport>>

pub async fn prepend_header( &self, name: Option<&str>, value: &str, ) -> Result<()>

pub async fn append_header(&self, name: Option<&str>, value: &str) -> Result<()>

pub async fn get_address_header( &self, header_name: &str, ) -> Result<Option<HeaderAddressList>>

pub async fn get_first_named_header_value( &self, name: &str, ) -> Result<Option<String>>

pub async fn get_all_named_header_values( &self, name: &str, ) -> Result<Vec<String>>

pub async fn get_all_headers(&self) -> Result<Vec<(String, String)>>

pub async fn retain_headers<F: FnMut(&Header<'_>) -> bool>( &self, func: F, ) -> Result<()>

pub async fn remove_first_named_header(&self, name: &str) -> Result<()>

pub async fn import_x_headers(&self, names: Vec<String>) -> Result<()>

pub async fn remove_x_headers(&self, names: Vec<String>) -> Result<()>

pub async fn remove_all_named_headers(&self, name: &str) -> Result<()>

pub async fn dkim_sign(&self, signer: Signer) -> Result<()>

pub async fn import_scheduling_header( &self, header_name: &str, remove: bool, ) -> Result<Option<Scheduling>>

pub async fn append_text_plain(&self, content: &str) -> Result<bool>

pub async fn append_text_html(&self, content: &str) -> Result<bool>

pub async fn check_fix_conformance( &self, check: MessageConformance, fix: MessageConformance, settings: Option<&CheckFixSettings>, ) -> Result<()>

impl Message

pub async fn serialize_for_xfer( &self, additional_meta: Value, ) -> Result<Vec<u8>>

pub fn deserialize_from_xfer(serialized: &[u8]) -> Result<Self>